+++ crates/scheduler/test.sh
@@ -31,5 +31,7 @@ cargo check --lib --target=i686-unknown-linux-gnu --features=full-api,native,std
 cargo check --lib --target=i686-unknown-linux-gnu --features=full-api,native,std,log
 cargo check --lib --target=thumbv7em-none-eabi --features=full-api,wasm
 DEFMT_LOG=trace cargo check --lib --target=thumbv7em-none-eabi --features=full-api,wasm,defmt
+cargo check --lib --target=thumbv7em-none-eabi --features=full-api,pulley
+DEFMT_LOG=trace cargo check --lib --target=thumbv7em-none-eabi --features=full-api,pulley,defmt
 cargo check --lib --target=thumbv7em-none-eabi --features=full-api,native
 DEFMT_LOG=trace cargo check --lib --target=thumbv7em-none-eabi --features=full-api,native,defmt